services/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
aio-pika==9.3.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sentry-sdk[fastapi]==1.40.0
PyJWT==2.8.0
cryptography==41.0.7
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
weaviate-client>=4.0.0
google-genai>=0.1.0
python-dateutil==2.9.0
bcrypt==4.1.2
email-validator==2.1.0
python-multipart==0.0.6
assemblyai
scikit-learn>=1.3.0
umap-learn>=0.5.0
numpy>=1.24.0
